DAY_13/requirements.txt
streamlit
google-generativeai
SpeechRecognition
pyttsx3
python-dotenv
